add boost fallback for DeleteFile
