+++ CommandLine/libEGM/filesystem.h
@@ -1,0 +1,25 @@
+#ifdef USE_BOOST_FS
+  #include <boost/filesystem.hpp>
+  #include <boost/system/error_code.hpp>
+  namespace fs = boost::filesystem;
+  using errc = boost::system::error_code;
+  namespace boost {
+    template<class ForwardIt>
+    ForwardIt next(
+        ForwardIt it,
+        typename std::iterator_traits<ForwardIt>::difference_type n = 1);
+  }  // namespace boost
+#else
+  #include <filesystem>
+  namespace fs = std::filesystem;
+  using errc = std::error_code;
+#endif
+
+using std::string;
+
+bool StartsWith(const string &str, const string &prefix);
+string StripPath(string fName);
+bool CreateDirectory(const fs::path &directory);
+fs::path InternalizeFile(const fs::path &file,
+                         const fs::path &directory, const fs::path &egm_root);
+void DeleteFile(const string &fName);
